Enhanced ICU driver.
git-svn-id: svn://svn.code.sf.net/p/chibios/svn/trunk@7211 35acf78f-673a-0410-8e92-d51de3d6d3f4

+++ os/hal/ports/STM32/LLD/TIMv1/pwm_lld.c
@@ -109,7 +109,7 @@ PWMDriver PWMD9;
  * @param[in] pwmp      pointer to a @p PWMDriver object
  */
 static void pwm_lld_serve_interrupt(PWMDriver *pwmp) {
-  uint16_t sr;
+  uint32_t sr;
 
   sr  = pwmp->tim->SR;
   sr &= pwmp->tim->DIER & STM32_TIM_DIER_IRQ_MASK;
